+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Fix gradle deprecation warning (#67)
 * Skip check task when no signatures configured (to differentiate with success execution 
   and easily spot configuration problems)
+* Add debug output for check and build tasks (to simplify configuration debugging)
+  - animalsniffer.debug = true for check tasks
+  - animalsnifferSignatire.debug = true for signature build task
 
 ### 1.6.0 (2022-08-20)
 * Update animalsniffer 1.20 -> 1.22 (java 9 support)